@@ -2,7 +2,7 @@ name: tripit
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: ">=3.3.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -31,22 +31,24 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
   firebase_core: ^2.27.2
-  google_fonts: 4.0.4
+  google_fonts: ^6.2.0
   pdf: ^3.10.8
   google_maps_flutter: ^2.6.0
   cloud_firestore: ^4.15.10
   intl: ^0.19.0
   geolocator: ^11.0.0
   firebase_auth: ^4.19.1
-  flutter_google_places: ^0.3.0
-  google_maps_webservice: ^0.0.20-nullsafety.5
+  #flutter_google_places: 0.2.7
+  # google_maps_webservice: ^0.0.20-nullsafety.0
   google_api_headers: ^1.6.0
   awesome_snackbar_content: ^0.1.1
+  google_generative_ai: ^0.1.0
+  firebase_database: ^10.1.1
+  webview_flutter: ^4.7.0
 
 dev_dependencies:
   flutter_test:
@@ -64,7 +66,6 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
