@@ -1,113 +1,113 @@
-import 'package:flutter/material.dart';
-import 'package:geolocator/geolocator.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:tripit/pages/search_places_page.dart'; // Import the SearchPlacesPage
+// import 'package:flutter/material.dart';
+// import 'package:geolocator/geolocator.dart';
+// import 'package:google_maps_flutter/google_maps_flutter.dart';
+// import 'package:tripit/pages/search_places_page.dart'; // Import the SearchPlacesPage
 
-class UserLocationPage extends StatefulWidget {
-  const UserLocationPage({super.key});
+// class UserLocationPage extends StatefulWidget {
+//   const UserLocationPage({super.key});
 
-  @override
-  State<UserLocationPage> createState() => _UserLocationPageState();
-}
+//   @override
+//   State<UserLocationPage> createState() => _UserLocationPageState();
+// }
 
-class _UserLocationPageState extends State<UserLocationPage> {
-  late GoogleMapController googleMapController;
+// class _UserLocationPageState extends State<UserLocationPage> {
+//   late GoogleMapController googleMapController;
 
-  static const CameraPosition initialCameraPosition =
-      CameraPosition(target: LatLng(6.9271, 79.8612), zoom: 14.0);
+//   static const CameraPosition initialCameraPosition =
+//       CameraPosition(target: LatLng(6.9271, 79.8612), zoom: 14.0);
 
-  Set<Marker> markers = {};
+//   Set<Marker> markers = {};
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.green,
-        title: const Text(
-          "Your Location",
-          style: TextStyle(color: Colors.white),
-        ),
-      ),
-      body: GoogleMap(
-        initialCameraPosition: initialCameraPosition,
-        markers: markers,
-        zoomControlsEnabled: false,
-        mapType: MapType.normal,
-        onMapCreated: (GoogleMapController controller) {
-          googleMapController = controller;
-        },
-      ),
-      floatingActionButton: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: [
-          FloatingActionButton.extended(
-            onPressed: () async {
-              Position position = await _determineposition();
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       appBar: AppBar(
+//         backgroundColor: Colors.green,
+//         title: const Text(
+//           "Your Location",
+//           style: TextStyle(color: Colors.white),
+//         ),
+//       ),
+//       body: GoogleMap(
+//         initialCameraPosition: initialCameraPosition,
+//         markers: markers,
+//         zoomControlsEnabled: false,
+//         mapType: MapType.normal,
+//         onMapCreated: (GoogleMapController controller) {
+//           googleMapController = controller;
+//         },
+//       ),
+//       floatingActionButton: Column(
+//         mainAxisAlignment: MainAxisAlignment.end,
+//         children: [
+//           FloatingActionButton.extended(
+//             onPressed: () async {
+//               Position position = await _determineposition();
 
-              googleMapController.animateCamera(
-                CameraUpdate.newCameraPosition(
-                  CameraPosition(
-                    target: LatLng(position.latitude, position.longitude),
-                    zoom: 14,
-                  ),
-                ),
-              );
+//               googleMapController.animateCamera(
+//                 CameraUpdate.newCameraPosition(
+//                   CameraPosition(
+//                     target: LatLng(position.latitude, position.longitude),
+//                     zoom: 14,
+//                   ),
+//                 ),
+//               );
 
-              markers.clear();
-              markers.add(
-                Marker(
-                  markerId: const MarkerId('currentLocation'),
-                  position: LatLng(position.latitude, position.longitude),
-                ),
-              );
+//               markers.clear();
+//               markers.add(
+//                 Marker(
+//                   markerId: const MarkerId('currentLocation'),
+//                   position: LatLng(position.latitude, position.longitude),
+//                 ),
+//               );
 
-              setState(() {});
-            },
-            label: const Text("Go To Your Location"),
-            icon: const Icon(Icons.location_history),
-          ),
-          SizedBox(height: 10), // Add some spacing between buttons
-          FloatingActionButton.extended(
-            onPressed: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => SearchPlacesPage()),
-              );
-            },
-            label: const Text("Search Places"),
-            icon: const Icon(Icons.search),
-          ),
-        ],
-      ),
-    );
-  }
+//               setState(() {});
+//             },
+//             label: const Text("Go To Your Location"),
+//             icon: const Icon(Icons.location_history),
+//           ),
+//           SizedBox(height: 10), // Add some spacing between buttons
+//           FloatingActionButton.extended(
+//             onPressed: () {
+//               Navigator.push(
+//                 context,
+//                 MaterialPageRoute(builder: (context) => SearchPlacesPage()),
+//               );
+//             },
+//             label: const Text("Search Places"),
+//             icon: const Icon(Icons.search),
+//           ),
+//         ],
+//       ),
+//     );
+//   }
 
-  Future<Position> _determineposition() async {
-    bool serviceEnabled;
-    LocationPermission permission;
+//   Future<Position> _determineposition() async {
+//     bool serviceEnabled;
+//     LocationPermission permission;
 
-    serviceEnabled = await Geolocator.isLocationServiceEnabled();
+//     serviceEnabled = await Geolocator.isLocationServiceEnabled();
 
-    if (!serviceEnabled) {
-      return Future.error("Location Services are disabled");
-    }
+//     if (!serviceEnabled) {
+//       return Future.error("Location Services are disabled");
+//     }
 
-    permission = await Geolocator.checkPermission();
+//     permission = await Geolocator.checkPermission();
 
-    if (permission == LocationPermission.denied) {
-      permission = await Geolocator.requestPermission();
+//     if (permission == LocationPermission.denied) {
+//       permission = await Geolocator.requestPermission();
 
-      if (permission == LocationPermission.denied) {
-        return Future.error("Location Permission denied");
-      }
-    }
+//       if (permission == LocationPermission.denied) {
+//         return Future.error("Location Permission denied");
+//       }
+//     }
 
-    if (permission == LocationPermission.deniedForever) {
-      return Future.error('Location Permissions are permenently denied');
-    }
+//     if (permission == LocationPermission.deniedForever) {
+//       return Future.error('Location Permissions are permenently denied');
+//     }
 
-    Position position = await Geolocator.getCurrentPosition();
+//     Position position = await Geolocator.getCurrentPosition();
 
-    return position;
-  }
-}
+//     return position;
+//   }
+// }
